Move a function and cleanup
