+++ website/src/pages/[organism]/search/index.astro
@@ -1,13 +1,18 @@
 ---
 import { cleanOrganism } from '../../../components/Navigation/cleanOrganism';
 import { SearchFullUI } from '../../../components/SearchPage/SearchFullUI';
-import { dataUseTermsAreEnabled, getRuntimeConfig, getSchema, getWebsiteConfig } from '../../../config';
+import {
+    dataUseTermsAreEnabled,
+    getReferenceGenomesSequenceNames,
+    getRuntimeConfig,
+    getSchema,
+    getWebsiteConfig,
+} from '../../../config';
 import BaseLayout from '../../../layouts/BaseLayout.astro';
 import { VERSION_STATUS_FIELD, IS_REVOCATION_FIELD } from '../../../settings';
 import { versionStatuses } from '../../../types/lapis';
 import { getAccessToken } from '../../../utils/getAccessToken';
 import { getMyGroups } from '../../../utils/getMyGroups';
-import { getReferenceGenomesSequenceNames } from '../../../utils/search';
 import { performLapisSearchQueries } from '../../../utils/serversideSearch';
 
 const hiddenFieldValues = {
