+++ website/src/pages/seq/[accessionVersion]/index.astro
@@ -6,13 +6,12 @@ import { SequenceDataUI } from '../../../components/SequenceDetailsPage/Sequence
 import SequencesBanner from '../../../components/SequenceDetailsPage/SequencesBanner.tsx';
 import SequencesDataTableTitle from '../../../components/SequenceDetailsPage/SequencesDataTableTitle.astro';
 import ErrorBox from '../../../components/common/ErrorBox.tsx';
-import { getSchema, getRuntimeConfig } from '../../../config';
+import { getSchema, getRuntimeConfig, getReferenceGenomesSequenceNames } from '../../../config';
 import { getWebsiteConfig } from '../../../config';
 import BaseLayout from '../../../layouts/BaseLayout.astro';
 import { type Group } from '../../../types/backend';
 import { getAccessToken } from '../../../utils/getAccessToken';
 import { getMyGroups } from '../../../utils/getMyGroups';
-import { getReferenceGenomesSequenceNames } from '../../../utils/search';
 
 const accessionVersion = Astro.params.accessionVersion!;
